# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/{id}/block": {
      "put": {
        "operationId": "ModerationController_block",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "moderation"
        ]
      },
      "delete": {
        "operationId": "ModerationController_unblock",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "moderation"
        ]
      }
    },
    "/users/{id}/report": {
      "post": {
        "operationId": "ModerationController_report",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string",
                    "enum": [
                      "HARASSMENT",
                      "INAPPROPRIATE_CONTENT",
                      "CHEATING",
                      "NO_SHOW",
                      "MINOR_SAFETY",
                      "OTHER"
                    ]
                  },
                  "details": {
                    "default": "",
                    "type": "string",
                    "maxLength": 1000
                  }
                },
                "required": [
                  "reason"
                ]
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "moderation"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Kwatro",
    "description": "",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "dev-user": {
        "type": "apiKey",
        "in": "header",
        "name": "x-dev-user-id"
      }
    },
...
  },
  "security": [
    {
      "dev-user": []
    }
  ]
}